take opportunity to consolidate tc_host code
